rename: supabase.ts -> client.ts
